requirements.txt
contourpy==1.1.1
cycler==0.12.1
fonttools==4.47.0
importlib-resources==6.1.1
keras==2.15.0
kiwisolver==1.4.5
matplotlib==3.7.4
numpy==1.24.4
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
scipy==1.10.1
six==1.16.0
zipp==3.17.0
